Run training as a command via cron
This patchset introduces a `cmd`-container, which runs the entire
training-pipeline via a cronjob.

The pipeline consists of fetching data, pre-processing, building
the model via KGAT, clustering, and post-processing.
Final data is stored in a shared volume, accessible by the `server`-
container, which serves this data publicly.

There are a few hacks involved in building the `cmd`-image, most
notably provisioning the `py3nctx`-module. The boost-libs required for
compiling this module pull in a python3.10 dependency that would clash
with the python3.11 included in the tensorflow-image.
To get around this, we use a build-stage, and copy the .so-files for boost
into the final image, gluing everything together via `ldconfig` to get the
dynamic loading right.